fix impl to correctly implement signed integers
the original impl followed the rust example and used unsigned
integers.  however, the spec states that integers are signed.
now the generated code and the runtime are updated to support
signed integers.

+++ go/util.go
@@ -2,19 +2,19 @@ package snowball
 
 import "unicode/utf8"
 
-func splitAt(str string, mid uint) (string, string) {
+func splitAt(str string, mid int) (string, string) {
 	return str[:mid], str[mid:]
 }
 
-func min(a, b uint) uint {
+func min(a, b int) int {
 	if a < b {
 		return a
 	}
 	return b
 }
 
-func onCharBoundary(s string, pos uint) bool {
-	if pos >= uint(len(s)) {
+func onCharBoundary(s string, pos int) bool {
+	if pos >= len(s) {
 		return true
 	}
 	return utf8.RuneStart(s[pos])
@@ -23,6 +23,6 @@ func onCharBoundary(s string, pos uint) bool {
 // RuneCountInString is a wrapper around utf8.RuneCountInString
 // this allows us to not have to conditionally include
 // the utf8 package into some stemmers and not others
-func RuneCountInString(str string) uint {
-	return uint(utf8.RuneCountInString(str))
+func RuneCountInString(str string) int {
+	return utf8.RuneCountInString(str)
 }
